feat: Add N+4 escalation level
This level should never be achieved under normal circumstances. A new escalation level has been added to the `GetEscalationDescriptions` method in the `Utilities.cs` file. This new level is called "☢️ Nuclear ☢️" and represents users who are extremely dangerous to any community. The commit also includes an update to the method's switch statement to handle this new escalation level appropriately.